refactoring utils to be seperate from rest of modules
