+++ NFACT/NFACT_config/nfact_config_functions.py
@@ -3,7 +3,7 @@ from sklearn.decomposition import FastICA, NMF
 import json
 import os
 import argparse
-from NFACT.NFACT_decomp.utils.utils import colours
+from NFACT.utils.utils import colours
 
 
 def nfact_config_args():
