+++ NFACT/NFACT_decomp/regression/dual_regression.py
@@ -1,6 +1,6 @@
 from NFACT.NFACT_decomp.decomposition.decomp import normalise_components
 from NFACT.NFACT_decomp.decomposition.matrix_handling import load_fdt_matrix
-from NFACT.NFACT_decomp.utils.utils import error_and_exit
+from NFACT.utils.utils import error_and_exit
 from NFACT.NFACT_decomp.pipes.image_handling import save_dual_regression_images
 
 import numpy as np
